requirements.txt
streamlit
tensorflow==2.16.1
numpy
pillow
matplotlib
pandas
